requirements.txt
# Discord Bot
discord.py>=2.3.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation & Type Hints
pydantic>=2.0.0

# LLM & AI Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.1
langgraph>=0.0.1

# Vector Database & RAG
chromadb>=0.4.0
langchain-chroma>=0.1.0

# Web Server
flask>=2.3.0

# Jupyter Support (for notebooks)
ipython>=8.0.0
jupyter>=1.0.0
